requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
pybithumb==1.0.17 